feat: 결제 연동 후 렌트 호출

+++ Drtaa-Android/core-model/src/main/java/com/drtaa/core_model/data/PaymentCompletionInfo.kt
@@ -13,7 +13,7 @@ data class PaymentCompletionInfo(
     val userId: String,
     val paymentMethod: String,
     val purchasedAt: Date,
-    val carId: Long = 1,
+    val carId: Int = 1,
     val headCount: Int = 1,
     val rentStartTime: Date = Date(),
     val rentEndTime: Date = Date(System.currentTimeMillis() + HOUR_TO_MILLIS.value) // 1시간 후
